requirements.txt
# Python development environment dependencies

# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
requests>=2.25.0

# Development dependencies
pytest>=6.2.0
pytest-cov>=2.12.0
black>=21.0.0
flake8>=3.9.0
mypy>=0.910

# Documentation
sphinx>=4.0.0
sphinx-rtd-theme>=0.5.0

# Virtual environment
virtualenv>=20.0.0

# Codd development
codd-dev

# Gemini API
google-genai>=0.3.0

# Redis
redis>=5.0.0

# PyYAML
pyyaml>=6.0.0

# OpenAI
openai>=2.32.0